requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.2
pandas>=2.0.0
pika>=1.3.1
pytest>=7.3.1
httpx>=0.24.0
sentence-transformers>=2.2.2
python-multipart>=0.0.6
tqdm>=4.65.0
joblib>=1.2.0
langchain>=0.0.200
langchain-ollama>=0.0.1
langchain-community>=0.0.1
langchain-core>=0.0.1
sqlalchemy>=2.0.0
sqlitedict>=2.1.0 